Refactor repositories to use Spring Data JPA
- Removed IUserRepository interface and implemented UserRepository using JpaRepository.
- Updated InventoryRepository, OrderItemRepository, OrderRepository, ProductRepository, and ReviewRepository to extend JpaRepository and added necessary query methods.
- Replaced custom update and delete methods with JpaRepository's save and deleteById methods in UserServiceImpl, OrderServiceImpl, ReviewServiceImpl, and other service implementations.
- Enhanced ProductRepository with additional filtering and searching capabilities.
- Updated application properties for JPA and SQL logging settings.

+++ src/main/java/com/shopjoy/service/PerformanceComparisonService.java
@@ -124,7 +124,7 @@ public class PerformanceComparisonService {
             product.setProductId(i);
             product.setProductName("Product_" + i);
             product.setDescription("Description for product " + i);
-            product.setPrice(BigDecimal.valueOf(10 + random.nextDouble() * 990).doubleValue());
+            product.setPrice(BigDecimal.valueOf(10 + random.nextDouble() * 990));
             product.setCategoryId(1 + random.nextInt(10));
             product.setCreatedAt(LocalDateTime.now().minusDays(random.nextInt(365)));
             products.add(product);
